Split query editor into multiple components and helper files
